refactor: restructure feature folders
