+++ src/admin/github/client.ts
@@ -1,5 +1,5 @@
 import { Octokit } from "@octokit/core";
-import { createAuth } from "../../../../auth";
+import { createAuth } from "@config/auth";
 
 const auth = createAuth();
