+++ src/app/components/DarkModeManager.astro
