+++ src/app/layouts/GlobalLayout.astro
@@ -1,10 +1,10 @@
 ---
-import "../_styles/global.css";
+import "../styles/global.css";
 
 import { GOOGLE_ANALYTICS_ID } from "astro:env/client";
 import { ClientRouter } from "astro:transitions";
-import DarkModeManager from "../_components/DarkModeManager.astro";
-import GoogleAnalytics from "../_components/GoogleAnalytics.astro";
+import DarkModeManager from "../components/DarkModeManager.astro";
+import GoogleAnalytics from "../components/GoogleAnalytics.astro";
 
 type Props = {
 	title?: string;
