+++ src/blog/layouts/PostLayout.astro
@@ -1,6 +1,6 @@
 ---
 import { type CollectionEntry, render } from "astro:content";
-import PostShell from "../_components/PostShell.astro";
+import PostShell from "../ui/PostShell.astro";
 
 export interface Props {
 	prevPost: CollectionEntry<"blog"> | null;
