+++ src/blog/ui/PostCard.astro
@@ -1,5 +1,5 @@
 ---
-import { BORDER_COLORS, type VitaColor } from "../../../constants";
+import { BORDER_COLORS, type VitaColor } from "@shared/constants";
 
 type Props = {
 	id: string;
