+++ src/blog/ui/PostHeader.astro
@@ -1,7 +1,7 @@
 ---
 import { Icon } from "astro-icon/components";
-import { BG_COLORS, VitaColor } from "../../../constants";
-import DarkModeButton from "../../_components/DarkModeButton.astro";
+import { BG_COLORS, VitaColor } from "@shared/constants";
+import DarkModeButton from "@app/components/DarkModeButton.astro";
 
 export interface Props {
 	color: VitaColor;
