+++ src/blog/ui/PostPrevNext.astro
@@ -1,7 +1,7 @@
 ---
 import type { CollectionEntry } from "astro:content";
 import { Icon } from "astro-icon/components";
-import { BORDER_COLORS } from "../../../constants";
+import { BORDER_COLORS } from "@shared/constants";
 
 export interface Props {
 	originalPost: CollectionEntry<"blog">;
