+++ src/blog/ui/TopCard.astro
@@ -4,8 +4,8 @@ import {
 	SEARCH_ICON_TRANSITION_NAME,
 	SITE_HEADER_TRANSITION_NAME,
 	TITLE_TRANSITION_NAME,
-} from "../../../constants";
-import DarkModeButton from "../../_components/DarkModeButton.astro";
+} from "@shared/constants";
+import DarkModeButton from "@app/components/DarkModeButton.astro";
 import Card from "./Card.astro";
 ---
