+++ src/pages/[slug]/index.astro
@@ -1,6 +1,6 @@
 ---
 import { type CollectionEntry, getCollection } from "astro:content";
-import PostLayout from "../../features/posts/_layouts/PostLayout.astro";
+import PostLayout from "@blog/layouts/PostLayout.astro";
 
 export interface Props {
 	post: CollectionEntry<"blog">;
