+++ src/pages/[slug]/og-image.png.ts
@@ -4,7 +4,7 @@ import { Resvg } from "@resvg/resvg-js";
 import type { APIRoute } from "astro";
 import { createElement } from "react";
 import satori from "satori";
-import { OGImage } from "../../features/posts/_components/OGImage";
+import { OGImage } from "@blog/ui/OGImage";
 
 export const GET: APIRoute = async ({ params }) => {
 	const { slug } = params;
