+++ src/pages/admin/edit/[slug].astro
@@ -1,8 +1,8 @@
 ---
-import GlobalLayout from "../../../features/_layouts/GlobalLayout.astro";
-import EditPost from "../../../features/admin/editor/_components/EditPost.astro";
-import { loadEditablePost } from "../../../features/admin/editor/load-editable-post";
-import type { EditablePost } from "../../../features/admin/editor/types";
+import GlobalLayout from "@app/layouts/GlobalLayout.astro";
+import EditPost from "@admin/editor/_components/EditPost.astro";
+import { loadEditablePost } from "@admin/editor/load-editable-post";
+import type { EditablePost } from "@admin/editor/types";
 
 export const prerender = false;
