+++ src/pages/admin/preview/[slug].astro
@@ -1,7 +1,7 @@
 ---
 import { getLiveEntry } from "astro:content";
-import { getGitHubAccessToken } from "../../../features/admin/_lib/github/client";
-import PostShell from "../../../features/posts/_components/PostShell.astro";
+import { getGitHubAccessToken } from "@admin/github/client";
+import PostShell from "@blog/ui/PostShell.astro";
 
 export const prerender = false;
