+++ src/pages/login.astro
@@ -1,5 +1,5 @@
 ---
-import GlobalLayout from "../features/_layouts/GlobalLayout.astro";
+import GlobalLayout from "@app/layouts/GlobalLayout.astro";
 
 export const prerender = false;
 ---
@@ -11,9 +11,7 @@ export const prerender = false;
 	</div>
 
 	<script>
-		const { signIn, signOut } = await import(
-			"../features/admin/_lib/auth-client"
-		);
+		const { signIn, signOut } = await import("@admin/auth/auth-client");
 		const loginButton = document.querySelector("#login");
 		const logoutButton = document.querySelector("#logout");
