+++ src/pages/search.astro
@@ -1,7 +1,7 @@
 ---
-import GlobalLayout from "../features/_layouts/GlobalLayout.astro";
-import SearchForm from "../features/search/_components/SearchForm.astro";
-import SearchResult from "../features/search/_components/SearchResult.astro";
+import GlobalLayout from "@app/layouts/GlobalLayout.astro";
+import SearchForm from "@search/ui/SearchForm.astro";
+import SearchResult from "@search/ui/SearchResult.astro";
 ---
 
 <GlobalLayout title="Search">
@@ -14,7 +14,7 @@ import SearchResult from "../features/search/_components/SearchResult.astro";
 </GlobalLayout>
 
 <script>
-	import { initSearch } from "../features/search/client";
+	import { initSearch } from "@search/client";
 	document.addEventListener("astro:after-swap", initSearch);
 	initSearch();
 </script>
