+++ src/shared/types/global.d.ts
