+++ src/shared/types/pagefind.d.ts
