+++ src/shared/utils/color.ts
@@ -1,0 +1,5 @@
+import type { VitaColor } from "../constants";
+
+export const getColorIndex = (date: Date): VitaColor => {
+	return (date.getDate() % 7) as VitaColor;
+};
